Instead of carping errors, sets the $error parameter.
Correctly set the y-axis label to the type of plot.

Simplified some R commands by setting any parameters which just require a
scalar using the perl scalar directly rather than setting an intermediate R
variable from the perl scalar.
Wrapped the call to R->run in an eval{} so program does not crash on an R
error.

Checks output from R and if it is a 'null device' error, ignores it as this
is not an error but is a message which is always returned when dev.off() is
called. If any other output is returned, this is considered and error and
returned to the caller via the $error parameter.